Provide another solution of utility connection for psql (--utility-mode option).
We have had the below solution for that but that is not convenient since the
command line is long sometimes.

	PGOPTIONS='-c gp_session_role=utility' psql

This patch provides another way of setting the psql connection in the utility
mode. The internal implementation is same as before (add the session role
guc setting in PGOPTIONS). Some code in the repo that still use the previous
solution is not modified.