Update minRecoveryPoint if needed during replay of XLOG_XACT_DISTRIBUTED_COMMIT
xlog XLOG_XACT_DISTRIBUTED_COMMIT is used on master to do 2pc commit, and
similar to local transaction commit, there might be dropping of relation files
or database directories when replaying the xlog.  Previously we do not update
minimum recovery point when doing that.  Minimum recovery point is used to
check consistency, when we replay to that point we know the database is
consistent and then so-called invalid-pages during replay are not allowed. On
upstream hot standby is allowed when it reaches consistency. Greenplum does not
support that but we might support that later.