Skip the pg data directory fsync step in startup after gprecoverseg.
gprecoverseg calls pg_basebackup or pg_rewind to do recovery on the failed
nodes and we force the pg data directory fsync in those commands so we could
skip the pg data directory fsync during startup since the recovered database
is not in cleanly shutdown states.

This could save some time if the recovered databases include a lot of table
files. It's implemented by writing a new file recovery_sync_done in the pg data
directory after the pg_basebackup/pg_rewind command succeeds. The file is
unlinked during startup.