Delay promote a bit in fts probe until the startup process starts recovery.
This tries to fix the below error in real environment:

could not rename file ""pg_replslot/internal_wal_replication_slot.tmp"" to
""pg_replslot/internal_wal_replication_slot"": File exists"

On greenplum, internal wal replication slot is created on primaries and is
dropped in mirrors during recovery - this often happens during gprecoverseg
rebalance. At that time the startup process is doing fsync on the whole pg data
directory after crash so if the replication slot exists it has not been dropped
yet. However I cannot reproduce this based on this theory after short try since
at that time fts does not promote due to 'double failure' - primary is down
and it thinks the streaming is not synchronized. I suspect there is some
corner cases I missed. Although the error might be not serious for the scenario
in theory, it would be better to the code to avoid such error.